Removed the idea of a commaDelimeterNode, changed atoi to atof

+++ Assignment2KML/kml.lex.cpp
@@ -950,7 +950,7 @@ case 9:
 YY_RULE_SETUP
 #line 128 "kml.l"
 {
-  yylval = node(atoi(yytext));
+  yylval = node(atof(yytext));
     #if DEBUG_LEXER == 1
       std::cout << "lexed " << yytext << " as NUMBER_LITERAL: " << yylval << std::endl;
     #endif
